Add a schema for stack trace frames.
