+++ packages/qmi/src/frame.tsx
@@ -1,15 +1,7 @@
+import {
+	FrameItem,
+} from 'qmi/data-types';
 import * as React from 'react';
-
-export interface FrameItem {
-	display: string;
-	args: string[];
-	calling_file: string;
-	calling_line: number;
-	file: string;
-	function: string;
-	id: string;
-	line: number;
-}
 
 export interface FrameProps {
 	frame: FrameItem;
