+++ src/schemas/data/logger.json
@@ -31,7 +31,14 @@
 						"type": "string"
 					},
 					"filtered_trace": {
-						"type": "array"
+						"type": "array",
+						"items": {
+							"allOf": [
+								{
+									"$ref": "../frameitem.json"
+								}
+							]
+						}
 					},
 					"component": {
 						"phpType": "QM_Component",
